requirements.txt
django>=4.2,<5.0
djangorestframework>=3.16
psycopg[binary]>=3.1
django-redis>=6.0
redis>=6.2.0
celery>=5.4.0
django-environ>=0.12.0
gunicorn>=23.0.0
pytest>=8.4.1
pytest-django>=4.11.1
ruff>=0.12.3
pillow>=11.3.0
drf-spectacular>=0.28.0
django-filter>=25.1
aiohttp>=3.9.5
